refactor: split out logger/tasks/output/spinners into package
