+++ packages/@ionic/cli-framework-output/jest.config.js
@@ -1,0 +1,1 @@
+module.exports = require('../../../jest.config.base');
