+++ packages/@ionic/cli-framework-output/lint-staged.config.js
@@ -1,0 +1,1 @@
+module.exports = require('../../../lint-staged.config.base');
