+++ packages/@ionic/cli-framework-output/package.json
@@ -1,0 +1,46 @@
+{
+  "name": "@ionic/cli-framework-output",
+  "version": "1.0.0",
+  "description": "The log/tasks/spinners portion of Ionic CLI Framework",
+  "homepage": "https://ionicframework.com/",
+  "author": "Ionic Team <hi@ionic.io> (https://ionicframework.com)",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "engines": {
+    "node": ">=10.3.0"
+  },
+  "files": [
+    "dist/",
+    "LICENSE",
+    "README.md"
+  ],
+  "scripts": {
+    "clean": "rimraf dist",
+    "lint": "tslint --project tsconfig.json",
+    "build": "npm run clean && tsc",
+    "watch": "tsc -w --preserveWatchOutput",
+    "test": "jest --maxWorkers=4",
+    "prepublishOnly": "npm run build"
+  },
+  "license": "MIT",
+  "dependencies": {
+    "@ionic/utils-terminal": "2.1.3",
+    "debug": "^4.0.0",
+    "inquirer": "^7.0.0",
+    "log-update": "^4.0.0",
+    "tslib": "1.11.2"
+  },
+  "devDependencies": {
+    "@ionic/utils-stream": "3.1.3",
+    "@types/debug": "^4.1.1",
+    "@types/inquirer": "0.0.43",
+    "@types/jest": "^25.1.0",
+    "@types/node": "~10.17.13",
+    "jest": "^25.1.0",
+    "jest-cli": "^26.0.1",
+    "lint-staged": "^10.0.2",
+    "ts-jest": "~25.4.0",
+    "tslint": "^5.9.1",
+    "typescript": "~3.8.2"
+  }
+}
