+++ packages/@ionic/cli-framework-output/src/index.ts
@@ -1,0 +1,4 @@
+export * from './colors';
+export * from './logger';
+export * from './output';
+export * from './tasks';
