+++ packages/@ionic/cli-framework/package.json
@@ -19,6 +19,7 @@
   },
   "license": "MIT",
   "dependencies": {
+    "@ionic/cli-framework-output": "1.0.0",
     "@ionic/utils-array": "2.1.3",
     "@ionic/utils-fs": "3.1.3",
     "@ionic/utils-object": "2.1.3",
@@ -29,7 +30,6 @@
     "chalk": "^4.0.0",
     "debug": "^4.0.0",
     "lodash": "^4.17.5",
-    "log-update": "^4.0.0",
     "minimist": "^1.2.0",
     "rimraf": "^3.0.0",
     "slice-ansi": "^4.0.0",
