+++ packages/@ionic/cli-framework/src/utils/__tests__/string.ts
@@ -1,4 +1,4 @@
-import { isValidURL, enforceLF, slugify } from '../string';
+import { isValidURL, slugify } from '../string';
 
 describe('@ionic/cli-framework', () => {
 
@@ -33,35 +33,6 @@ describe('@ionic/cli-framework', () => {
       it('should test a git url', () => {
         expect(isValidURL('git@github.com:ionic-team/ionic-cli.git')).toBe(true);
       });
-    });
-
-    describe('enforceLF', () => {
-
-      it('should convert empty string to newline', () => {
-        const result = enforceLF('');
-        expect(result).toBe('\n');
-      });
-
-      it('should do nothing to a single newline', () => {
-        const result = enforceLF('\n');
-        expect(result).toBe('\n');
-      });
-
-      it('should add newline to text without newline', () => {
-        const result = enforceLF('some text');
-        expect(result).toBe('some text\n');
-      });
-
-      it('should do nothing to multiline text with newline at end', () => {
-        const result = enforceLF('some text\nsome more\n');
-        expect(result).toBe('some text\nsome more\n');
-      });
-
-      it('should add newline to multiline text without newline at end', () => {
-        const result = enforceLF('some text\nsome more');
-        expect(result).toBe('some text\nsome more\n');
-      });
-
     });
 
     describe('slugify', () => {
